# Faster IntSet and IntMap split functions