--- a/containers/src/Data/IntSet/Internal.hs
+++ b/containers/src/Data/IntSet/Internal.hs
@@ -882,11 +882,11 @@ spanAntitone predicate t =
 split :: Key -> IntSet -> (IntSet,IntSet)
 split x t =
   case t of
-      Bin _ m l r
+      Bin p m l r
           | m < 0 -> if x >= 0  -- handle negative numbers.
-                     then case go x l of (lt :*: gt) -> let !lt' = union lt r
+                     then case go x l of (lt :*: gt) -> let !lt' = bin p m lt r
                                                         in (lt', gt)
-                     else case go x r of (lt :*: gt) -> let !gt' = union gt l
+                     else case go x r of (lt :*: gt) -> let !gt' = bin p m l gt
                                                         in (lt, gt')
       _ -> case go x t of
           (lt :*: gt) -> (lt, gt)
@@ -894,9 +894,9 @@ split x t =
     go !x' t'@(Bin p m l r)
         | match x' p m = if zero x' m
                          then case go x' l of
-                             (lt :*: gt) -> lt :*: union gt r
+                             (lt :*: gt) -> lt :*: bin p m gt r
                          else case go x' r of
-                             (lt :*: gt) -> union lt l :*: gt
+                             (lt :*: gt) -> bin p m l lt :*: gt
         | otherwise   = if x' < p then (Nil :*: t')
                         else (t' :*: Nil)
     go x' t'@(Tip kx' bm)
@@ -913,21 +913,21 @@ split x t =
 splitMember :: Key -> IntSet -> (IntSet,Bool,IntSet)
 splitMember x t =
   case t of
-      Bin _ m l r | m < 0 -> if x >= 0
+      Bin p m l r | m < 0 -> if x >= 0
                              then case go x l of
-                                 (lt, fnd, gt) -> let !lt' = union lt r
+                                 (lt, fnd, gt) -> let !lt' = bin p m lt r
                                                   in (lt', fnd, gt)
                              else case go x r of
-                                 (lt, fnd, gt) -> let !gt' = union gt l
+                                 (lt, fnd, gt) -> let !gt' = bin p m l gt
                                                   in (lt, fnd, gt')
       _ -> go x t
   where
     go x' t'@(Bin p m l r)
         | match x' p m = if zero x' m
                          then case go x' l of
-                             (lt, fnd, gt) -> (lt, fnd, union gt r)
+                             (lt, fnd, gt) -> (lt, fnd, bin p m gt r)
                          else case go x' r of
-                             (lt, fnd, gt) -> (union lt l, fnd, gt)
+                             (lt, fnd, gt) -> (bin p m l lt, fnd, gt)
         | otherwise   = if x' < p then (Nil, False, t') else (t', False, Nil)
     go x' t'@(Tip kx' bm)
         | kx' > x'          = (Nil, False, t')