Following up from [#874](https://github.com/haskell/containers/pull/874#discussion_r1027248480).

The main changes are using `bin` everywhere and being strict in `IntSet.splitMember`. Also added one missing property test , and benchmarks to measure the change.

IntSet:
```
  split:dense:        OK (0.20s)
    46.4 ns ± 2.7 ns, 49% less than baseline
  split:sparse:       OK (0.18s)
    84.2 ns ± 5.5 ns, 41% less than baseline
  splitMember:dense:  OK (0.20s)
    46.3 ns ± 3.6 ns, 47% less than baseline
  splitMember:sparse: OK (0.17s)
    82.1 ns ± 5.4 ns, 45% less than baseline
```

IntMap:
```
  split:       OK (0.19s)
    82.3 ns ± 6.4 ns, 43% less than baseline
  splitLookup: OK (0.17s)
    84.5 ns ± 5.9 ns, 43% less than baseline
```