@meooow25 2022-11-27
Side note: Antitone functions run much slower than split functions. For example, `spanAntitone` takes 1.6-2.6x more time than `splitMember` on the current benchmarks. Inlining it makes it just as fast, but would we want to do that?