@treeowl 2022-11-27
Impressive results! Where did you add strictness in `splitMember`? I can't find it. I don't *think* we want to `INLINE` the general antitone functions, but it would be good to make sure that they get unfoldings and to document that inclining them might be a good idea in hot spots.