@meooow25 2022-11-27
I see. I'll check if that works separately.